オプションページのルーティングとテストを追加
