+++ components/mode-toggle.tsx
@@ -12,7 +12,7 @@ export const ModeToggle = () => {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild={true}>
-        <Button variant='outline' size='icon'>
+        <Button className='relative' variant='outline' size='icon'>
           <Sun className='h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0' />
           <Moon className='absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100' />
           <span className='sr-only'>テーマの切り替え</span>
